inet: fix inet_rtnl_recv to read multiple nlmsgs
When dumping multiple rules it's very likely that the reply will contain
multiple netlink messages (nlmsg) all with the same sequence number.

The original code was calling the callback only for the first message. This
patch is a refactoring of the function to use a clearer loop structure,
and calls the callback for every match IF the callback didn't return true.

A clearer example for this: Given the list of rules below we only want to
search for the second one in the list.

0:      from all lookup local
32764:  from 10.1.2.2 lookup 2
32765:  from 10.1.1.2 lookup 1
32766:  from all lookup main
32767:  from all lookup default

All rules will be returned and the callback will return true
only when it will get to the desired rule. The old behaviour would stop at
the first rule.