service: add/del multipath service route tables
Multipath tcp needs a routing table for every interface that uses it.
This routing table is configured to do source routing.

Always add source routing tables when an service enters READY state.
And always delete that table on DISCONNECT/FAILURE or on service
cleanup.

The table id is handled in the following way:
* Table id is derived from interface index + 255.
* IPv4 and IPv6 will use the same table id.

The table gets deleted when the user turns off multipath on that link.