service: correctly monitor links when triggering by timeout
There has been a logic error in the link monitoring, that the WISPr
auth is not done even when monitor_timeout has expired. To fix that,
move a call to __connman_wispr_start() to the timeout callback,
monitor_timeout_triggered(). Add also a missing call to
set_monitor_timeout() to be able to go again into the monitoring loop.