service: fix re-add or del of empty mpath table
Multipath routing tables were added every time multipath_add_routing_table
was called. We only want to add it when ipconfig doesn't have a table set.

Table 0 was being deleted. But for us 0 is invalid. This is a result of
not checking if table_id is set in ipconfig before going on with deletion.