session: add MultipathRoutedService config
Allow setting a list of services via the dbus api that can do multipath
routing for this session policy. This practically means that for each
service added, a new source + fwmark rule that goes to the table of
the service multipath gets added. This is true only if the service is
in a connected state with multipath enabled.

There's no way to specifically add or delete services. To do so you
must set the list again with all the services that you want.