session: fix service+session rule deletion
Don't specify fwmark when deleting a session rule. Service
and session rules (for a fixed multipath service) use the
same routing table. That means that on disconnect, the service
code already deleted the session rule because it's the first
match. Now we need to delete the remaining rule. Here's an example:

32764:  from 10.10.3.2 fwmark 0x100 lookup 314 -> session rule
32765:  from 10.10.3.2 lookup 314 -> service rule

Here's what happens:
* service goes down:
  - ip rule delete from 10.10.3.2 table 314
  - pref rule 32764 gets deleted
* session gets notified of service being down:
  - ip rule delete from 10.10.3.2 fwmark 0x100 table 314
  - no match, 32765 remains there

There are 2 other possible solutions:
* Keep track of rule pref numbers. Involves adding lots of code to
read that.
* Keep something like a reference counter for every session and
on service down execute the rule deletion refcnt+1 times. I don't
like this approach because it is to be harder to debug if any issue
happens.