iptables: correctly set proto
The problem: Using tcp/multiport extensions would trigger errors
when trying to pass the rule to the kernel. The reason is
the following: When the protocol is parsed, it's set only in the
context local fw ipt_entry in parse_xt_modules. This makes the
userspace check to be ok, but the kernel check to fail because
another ipt_entry is passed on to the kernel (the one in parse_ctx).

The solution: Always use the ipt_entry from parse_ctx. And set the
protocol value there instead of parse_ctx.